+++ libheif/context.h
@@ -187,7 +187,7 @@ public:
   // If 0 is passed as track_id, the main visual track is returned (we assume that there is only one visual track).
   Result<std::shared_ptr<Track>> get_track(uint32_t track_id);
 
-  uint64_t get_sequence_timescale() const;
+  uint32_t get_sequence_timescale() const;
 
   uint64_t get_sequence_duration() const;
